dataflow/requirements.txt
kafka-python
requests
python-dotenv
Flask
pymysql